+++ src/components/button/social/apple/social-apple.stories.tsx
@@ -1,0 +1,28 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import SocialAppleButton from '.'
+
+const meta = {
+  title: 'Buttons/Social/Apple Social Button',
+  component: SocialAppleButton,
+  parameters: {
+    layout: 'centered',
+  },
+  tags: ['autodocs'],
+  args: {
+    size: 'md',
+    fullWidth: false,
+    hideLabel: false,
+  },
+  decorators: [
+    (Story) => (
+      <div className="flex min-h-[256px] min-w-[512px] items-center justify-center">
+        <Story />
+      </div>
+    ),
+  ],
+} satisfies Meta<typeof SocialAppleButton>
+
+export default meta
+type Story = StoryObj<typeof meta>
+
+export const Default: Story = {}