+++ src/components/input/text-field/index.tsx
@@ -3,9 +3,11 @@ import Input from '../base'
 
 export type TextFieldProps = ComponentProps<typeof Input>
 
-const TextField = forwardRef<HTMLInputElement, TextFieldProps>((props, ref) => {
-  return <Input {...props} ref={ref} />
-})
+export const TextField = forwardRef<HTMLInputElement, TextFieldProps>(
+  (props, ref) => {
+    return <Input {...props} ref={ref} />
+  },
+)
 
 TextField.displayName = 'TextField'