+++ src/components/typography/code/index.tsx
@@ -1,25 +1,40 @@
+import { VariantProps, cva } from 'class-variance-authority'
 import qtMerge from 'qtMerge'
-import { forwardRef } from 'react'
-import { BodyTypographyProps, TextSize } from 'types'
+import { HTMLAttributes, forwardRef } from 'react'
+import { BodyTextProps, bodyTextCVA } from '../text/text.classes'
 
-export const codeTextSizeClassnames: Record<TextSize, string> = {
-  sm: 'text-code-sm space-y-paragraphSpacing-code-sm leading-code-sm',
-  md: 'text-code-md space-y-paragraphSpacing-code-md leading-code-md',
-  lg: 'text-code-lg space-y-paragraphSpacing-code-lg leading-code-lg',
-  xl: 'text-code-xl space-y-paragraphSpacing-code-xl leading-code-xl',
-}
+const codeTextSizeCVA = cva('', {
+  variants: {
+    size: {
+      sm: 'text-code-sm space-y-paragraphSpacing-code-sm leading-code-sm',
+      md: 'text-code-md space-y-paragraphSpacing-code-md leading-code-md',
+      lg: 'text-code-lg space-y-paragraphSpacing-code-lg leading-code-lg',
+      xl: 'text-code-xl space-y-paragraphSpacing-code-xl leading-code-xl',
+    },
+  },
+})
 
-export const CodeText = forwardRef<HTMLParagraphElement, BodyTypographyProps>(
-  ({ size = 'md', className, bold, italic, underline, ...rest }, ref) => {
+export type CodeTextSizeProps = VariantProps<typeof codeTextSizeCVA>
+
+export interface CodeTextProps
+  extends HTMLAttributes<HTMLElement>,
+    BodyTextProps,
+    CodeTextSizeProps {}
+
+export const CodeText = forwardRef<HTMLParagraphElement, CodeTextProps>(
+  ({ size = 'md', bold, italic, className, underline, ...rest }, ref) => {
     return (
       <code
         ref={ref}
         className={qtMerge(
           'font-mono',
-          codeTextSizeClassnames[size],
-          bold ? 'font-bold' : 'font-regular',
-          italic ? 'italic' : 'not-italic',
-          underline && 'underline',
+          codeTextSizeCVA({ size }),
+          bodyTextCVA({
+            colorStyle: 'prominent',
+            bold,
+            italic,
+            underline,
+          }),
           className,
         )}
         {...rest}