+++ tailwind.config.js
@@ -5,7 +5,11 @@ export default {
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
   darkMode: 'class',
   theme: {
-    extend: {},
+    extend: {
+      colors: {
+        transparent: 'transparent',
+      },
+    },
   },
   plugins: [],
   presets: [quillTailwindConfig],